# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OneChat API specification
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface Profile {
  id: string;
  userId: string;
  displayName: string;
  /** @nullable */
  avatarUrl: string | null;
  icebreaker: string;
  /** @nullable */
  aura?: string | null;
  lastActive: string;
  createdAt: string;
}

export interface ProfileSetup {
  /**
     * @minLength 1
     * @maxLength 40
     */
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
  /**
     * @minLength 1
     * @maxLength 200
     */
  icebreaker: string;
  /** @nullable */
  aura?: string | null;
}

export interface ProfileUpdate {
  /**
     * @minLength 1
     * @maxLength 40
     */
  displayName?: string;
  /** @nullable */
  avatarUrl?: string | null;
  /**
     * @minLength 1
     * @maxLength 200
     */
  icebreaker?: string;
  /** @nullable */
  aura?: string | null;
}

export interface MatchPartner {
  userId: string;
  displayName: string;
  /** @nullable */
  avatarUrl: string | null;
  icebreaker: string;
  /** @nullable */
  aura?: string | null;
  /** @nullable */
  lastActive?: string | null;
}

export type MatchStateStatus = typeof MatchStateStatus[keyof typeof MatchStateStatus];


export const MatchStateStatus = {
  active: 'active',
  waiting: 'waiting',
  ended: 'ended',
  blocked: 'blocked',
  no_profile: 'no_profile',
} as const;

export interface MatchState {
  status: MatchStateStatus;
  /** @nullable */
  matchId?: string | null;
  partner?: MatchPartner | null;
  /** @nullable */
  matchDate?: string | null;
  /** @nullable */
  expiresAt?: string | null;
}

export interface EndMatchInput {
  block: boolean;
}

export interface Message {
  id: string;
  matchId: string;
  senderId: string;
  content: string;
  /** @nullable */
  replyToId?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  readAt?: string | null;
  createdAt: string;
}

export interface MessageInput {
  /** @maxLength 2000 */
  content?: string;
  /** @nullable */
  replyToId?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface ArchivedMatch {
  matchId: string;
  matchDate: string;
  partnerName: string;
  /** @nullable */
  partnerAvatarUrl?: string | null;
  /** @nullable */
  partnerIcebreaker?: string | null;
  /** @nullable */
  partnerAura?: string | null;
  messageCount: number;
  status: string;
  /** @nullable */
  firstMessage?: string | null;
  /**
     * Duration of conversation in minutes (time between first and last message)
     * @nullable
     */
  conversationDuration?: number | null;
}

export interface ReactionSummary {
  emoji: string;
  count: number;
  byMe: boolean;
}

export interface ReactionInput {
  emoji: string;
}

export interface ReactionAdded {
  messageId: string;
  emoji: string;
  userId: string;
}

export type ReportInputReason = typeof ReportInputReason[keyof typeof ReportInputReason];


export const ReportInputReason = {
  spam: 'spam',
  harassment: 'harassment',
  inappropriate: 'inappropriate',
  other: 'other',
} as const;

export interface ReportInput {
  reportedUserId: string;
  /** @nullable */
  matchId?: string | null;
  reason: ReportInputReason;
  /**
     * @maxLength 1000
     * @nullable
     */
  details?: string | null;
}

export interface ReportCreated {
  id: string;
  success: boolean;
}

export interface BlockEntry {
  id: string;
  blockedId: string;
  createdAt: string;
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};

export type GetMatchArchiveParams = {
/**
 * Search by partner name, icebreaker, or message content
 */
q?: string;
/**
 * @minimum 1
 */
page?: number;
/**
 * @minimum 1
 * @maximum 50
 */
limit?: number;
};

export type GetMatchMessagesParams = {
/**
 * Return messages created before this message ID
 */
before?: string;
/**
 * @minimum 1
 * @maximum 50
 */
limit?: number;
};