# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: OneChat API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /matches/{matchId}/end:
    post:
      operationId: endMatch
      tags: [matches]
      summary: End the active match (and optionally block partner)
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EndMatchInput"
      responses:
        "200":
          description: Match ended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchState"
  /matches/archive:
    get:
      operationId: getMatchArchive
      tags: [matches]
      summary: Get past matches with search and pagination
      parameters:
        - name: q
          in: query
          required: false
          description: Search by partner name, icebreaker, or message content
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        "200":
          description: List of archived matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArchivedMatch"
  /matches/{matchId}/messages:
    get:
      operationId: getMatchMessages
      tags: [messages]
      summary: Get messages for a match (cursor pagination)
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
        - name: before
          in: query
          required: false
          description: Return messages created before this message ID
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 30
      responses:
        "200":
          description: List of messages (oldest first)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message in the active match
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "403":
          description: Match is not active or expired
  /matches/{matchId}/messages/read:
    post:
      operationId: markMessagesRead
      tags: [messages]
      summary: Mark messages as read (triggers read receipt to partner)
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Read event broadcast
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
...
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    MatchPartner:
      type: object
      required: [userId, displayName, avatarUrl, icebreaker]
      properties:
        userId:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        icebreaker:
          type: string
        aura:
          type: ["string", "null"]
        lastActive:
          type: ["string", "null"]
    MatchState:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [active, waiting, ended, blocked, no_profile]
        matchId:
          type: ["string", "null"]
        partner:
          oneOf:
            - $ref: "#/components/schemas/MatchPartner"
            - type: "null"
        matchDate:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
    EndMatchInput:
      type: object
      required: [block]
      properties:
        block:
          type: boolean
    Message:
      type: object
      required: [id, matchId, senderId, content, createdAt]
      properties:
        id:
          type: string
        matchId:
          type: string
        senderId:
          type: string
        content:
          type: string
        replyToId:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        readAt:
          type: ["string", "null"]
        createdAt:
          type: string
    MessageInput:
      type: object
      properties:
        content:
          type: string
          maxLength: 2000
        replyToId:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
    ArchivedMatch:
      type: object
      required: [matchId, matchDate, partnerName, messageCount, status]
      properties:
        matchId:
          type: string
        matchDate:
          type: string
        partnerName:
          type: string
        partnerAvatarUrl:
          type: ["string", "null"]
        partnerIcebreaker:
          type: ["string", "null"]
        partnerAura:
          type: ["string", "null"]
        messageCount:
          type: integer
        status:
          type: string
        firstMessage:
          type: ["string", "null"]
        conversationDuration:
          type: ["integer", "null"]
          description: Duration of conversation in minutes (time between first and last message)
...