# lib/api-zod/src/generated/types/matchState.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OneChat API specification
 * OpenAPI spec version: 0.2.0
 */
import type { MatchPartner } from './matchPartner';
import type { MatchStateStatus } from './matchStateStatus';

export interface MatchState {
  status: MatchStateStatus;
  /** @nullable */
  matchId?: string | null;
  partner?: MatchPartner | null;
  /** @nullable */
  matchDate?: string | null;
  /** @nullable */
  expiresAt?: string | null;
}