# lib/api-zod/src/generated/types/messageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OneChat API specification
 * OpenAPI spec version: 0.2.0
 */

export interface MessageInput {
  /** @maxLength 2000 */
  content?: string;
  /** @nullable */
  replyToId?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}